build: specify panic=abort
Specify panic=abort for dev and release profiles in the workspace. This
makes it easier to compile the project from source checkouts.
Unfortunately, this does not necessarily mean all users will get these
configurations applied. It is up to the developer to compile from the
workspace or pick a suitable profile manually.

It is unlikely that cargo will get a way to specify panic-strategies
per binary. Hence, we will have to live with this workaround for now.

Signed-off-by: David Rheinsberg <david@readahead.eu>