runtime11: work around ARM unwinding ABI
Define the ARM unwinding personality routines to fix linking on ARM.
These are usually defined by libgcc (or whatever unwinding library is
used). For unwinding to work, this needs to be pulled in.

Unfortunately, even with unwinding disable as we do, these symbols are
still pulled in by libcore and surprisingly not stripped properly during
linking. Hence, provide them as stubs for now and decide on a better
home later.

Signed-off-by: David Rheinsberg <david@readahead.eu>